move some components to shareTaskComponents directory
