+++ src/scenes/Dashboard/components/TaskItem.js
@@ -1,4 +1,4 @@
-import React, { useEffect, useRef, useState } from "react";
+import React, { useEffect, useState } from "react";
 import useMediaQuery from "@mui/material/useMediaQuery";
 import { alpha } from "@mui/material";
 import { Link, useHistory, useLocation } from "react-router-dom";
