+++ src/scenes/GuidedSetup/components/CallerDetails.tsx
@@ -3,7 +3,7 @@ import * as models from "../../../models";
 import { Stack, TextField, Typography } from "@mui/material";
 import React from "react";
 import { ContactForm } from "../../../components/ContactForm";
-import EstablishmentDetails from "../../../components/EstablishmentDetails";
+import EstablishmentDetails from "../../sharedTaskComponents/EstablishmentDetails";
 
 type CallerDetailType = {
     onChangeContact: (contact: any) => void;
