+++ src/scenes/ScheduledTasks/components/AdminAddScheduledTask.tsx
@@ -1,5 +1,5 @@
 import React from "react";
-import { Button, Divider, Stack, Typography } from "@mui/material";
+import { Button, Divider, Stack } from "@mui/material";
 import ScheduledTaskCallerDetails from "./ScheduledTaskCallerDetails";
 import * as models from "../../../models";
 import { PaddedPaper } from "../../../styles/common";
