+++ src/scenes/ScheduledTasks/components/ScheduledTaskOverviewSummary.tsx
@@ -2,9 +2,9 @@ import React from "react";
 import { useTheme } from "@mui/material/styles";
 import * as models from "../../../models";
 import { Box, Divider, Paper, Stack, Typography } from "@mui/material";
-import TaskDetailsEstablishment from "../../../components/TaskDetailsEstablishment";
-import RequesterContact from "../../../components/RequesterContact";
-import PrioritySelect from "../../../components/PrioritySelect";
+import TaskDetailsEstablishment from "../../sharedTaskComponents/TaskDetailsEstablishment";
+import RequesterContact from "../../sharedTaskComponents/RequesterContact";
+import PrioritySelect from "../../sharedTaskComponents/PrioritySelect";
 import { DataStore } from "aws-amplify";
 import { displayErrorNotification } from "../../../redux/notifications/NotificationsActions";
 import { useDispatch } from "react-redux";
