+++ src/scenes/Task/components/TaskDetailsPanel.js
@@ -1,8 +1,8 @@
 import React, { useEffect, useRef, useState } from "react";
-import TaskDetailsEstablishment from "../../../components/TaskDetailsEstablishment";
+import TaskDetailsEstablishment from "../../sharedTaskComponents/TaskDetailsEstablishment";
 import Typography from "@mui/material/Typography";
 import LabelItemPair from "../../../components/LabelItemPair";
-import PrioritySelect from "../../../components/PrioritySelect";
+import PrioritySelect from "../../sharedTaskComponents/PrioritySelect";
 import PropTypes from "prop-types";
 import TimePicker from "./TimePicker";
 import { Divider, Paper, Skeleton, Stack } from "@mui/material";
@@ -13,7 +13,7 @@ import { useDispatch, useSelector } from "react-redux";
 import { displayErrorNotification } from "../../../redux/notifications/NotificationsActions";
 import { dataStoreModelSyncedStatusSelector } from "../../../redux/Selectors";
 import GetError from "../../../ErrorComponents/GetError";
-import RequesterContact from "../../../components/RequesterContact";
+import RequesterContact from "../../sharedTaskComponents/RequesterContact";
 import { userRoles } from "../../../apiConsts";
 import { useAssignmentRole } from "../../../hooks/useAssignmentRole";
 import RiderResponsibilityDetail from "./RiderResponsibilityDetail";
