+++ src/scenes/Task/components/TaskOverview.js
@@ -7,10 +7,10 @@ import * as models from "../../../models";
 import useMediaQuery from "@mui/material/useMediaQuery";
 import { useTheme } from "@mui/material/styles";
 import { makeStyles } from "tss-react/mui";
-import DeliverableDetails from "../../../components/DeliverableDetails";
+import DeliverableDetails from "../../sharedTaskComponents/DeliverableDetails";
 import TaskActions from "./TaskActions";
 import { Hidden, Stack } from "@mui/material";
-import LocationDetailsPanel from "../../../components/LocationDetailsPanel";
+import LocationDetailsPanel from "../../sharedTaskComponents/LocationDetailsPanel";
 import TaskAssignmentsPanel from "./TaskAssignmentsPanel";
 import CommentsSection from "../../Comments/CommentsSection";
